+++ apps/api/test/application/use-cases/git/decide-bootstrap-plan.use-case.spec.ts
@@ -66,6 +66,7 @@ class UnreachableEngineClient implements ApiToEngineClient {
   reviseStory = () => this.boom();
   offerInfraHandoff = () => this.boom();
   reanalyzeSession = () => this.boom();
+  getPsychologistStatus = () => this.boom();
   runAnamnese = () => this.boom();
   invalidateInstructions = () => this.boom();
   requestRunnerTicket = () => this.boom();